apply suggested changes - move new exception to _exceptions - remove type from nutrient units (in both code and tests) - improve splitting of nutrients block - add check for singular in nutrients unit
