+++ tests/test_data/marleyspoon.de/marleyspoon.json
@@ -38,9 +38,9 @@
   ],
   "nutrients": {
     "calories": "842 calories",
-    "fatContent": "46.3 grams fat",
-    "carbohydrateContent": "74.2 grams carbohydrates",
-    "proteinContent": "31.4 grams protein"
+    "fatContent": "46.3 grams",
+    "carbohydrateContent": "74.2 grams",
+    "proteinContent": "31.4 grams"
   },
   "image": "https://s3-eu-west-1.amazonaws.com/images-marleyspoon-production/media/recipes/371109/main_photos/large/kasespatzle-ea2f2cc5a77168cc5395851a50a6dc4b.jpeg",
   "keywords": [
